Implement namei

+++ src/filesystem/params.rs
@@ -9,7 +9,7 @@ pub const NOFILE: usize =       16;     // open files per process
 pub const NFILE: usize =       100;     // open files per system
 pub const NINODE: usize =       50;     // maximum number of active i-nodes
 pub const NDEV: usize =         10;     // maximum major device number
-pub const ROOTDEV: usize =       1;     // device number of file system root disk
+pub const ROOTDEV: u32 =       1;     // device number of file system root disk
 pub const MAXARG: usize =       32;     // max exec arguments
 pub const MAXOPBLOCKS: usize =  10;     // max # of blocks any FS op writes
 pub const LOGSIZE: usize =      (MAXOPBLOCKS*3);  // max data blocks in on-disk log
@@ -19,7 +19,7 @@ pub const BPB: usize =          BSIZE * 8; // bits per block
 pub const FSSIZE: usize =       1000;   // size of file system in blocks
 
 // https://github.com/mit-pdos/xv6-riscv/blob/riscv/kernel/fs.h
-pub const ROOTINO: usize =      1;      // root i-number
+pub const ROOTINO: u32   =      1;      // root i-number
 pub const FSMAGIC: usize =      0x10203040;
 pub const NDIRECT: usize =      12;
 pub const NINDIRECT: usize =    BSIZE / mem::size_of::<u32>();
